ref(main): Move modals, loaders and other views to app component
