+++ src/pages/main/main.jsx
@@ -18,20 +18,11 @@ import styles from './main.module.css';
 export function MainPage() {
   const dispatch = useDispatch();
 
-  const [loadingText, setLoadingText] = useState('');
-
   const orderNumber = useSelector((state) => state.order.orderNumber);
   const detailedIngredient = useSelector((state) => state.currentIngredient.detailedIngredient);
 
   const ingredients = useSelector((state) => state.ingredients.ingredients);
   const isLoading = useSelector((state) => state.ingredients.loading);
-  const ingredientsErrorText = useSelector((state) => state.ingredients.error);
-  const orderErrorText = useSelector((state) => state.order.error);
-
-  const hasError = useCallback(
-    () => ingredientsErrorText || orderErrorText,
-    [ingredientsErrorText, orderErrorText],
-  );
 
   const handleCloseOrderModal = () => {
     dispatch({ type: Actions.RESET_CONSTRUCTOR_INGREDIENTS });
@@ -39,40 +30,9 @@ export function MainPage() {
   };
   const handleCloseDetailedIngredientModal = () => dispatch({ type: Actions.RESET_DETAILED_INGREDIENT });
 
-  useEffect(() => {
-    let interval; let
-      tick = 0;
-    if (isLoading) {
-      interval = setInterval(() => {
-        setLoadingText(`Верстаем меню${'.'.repeat(tick)}`);
-        tick = tick === 3 ? 0 : ++tick;
-      }, 450);
-    }
-
-    return () => clearInterval(interval);
-  }, [isLoading]);
-
-  useEffect(() => {
-    dispatch(getIngredients());
-  }, [dispatch]);
   return (
     <>
-      {isLoading && (
-        <div className={classNames(styles.loader, 'text', 'text_type_main-default')}>{loadingText}</div>
-      )}
-      {ingredientsErrorText && (
-        <div className={classNames(styles.error, 'text', 'text_type_main-default')}>
-          При загрузке ингредиетов произошла ошибка... 😞
-          <span className={classNames('mt-2', 'text', 'text_type_main-default', 'text_color_inactive')}>{ingredientsErrorText}</span>
-        </div>
-      )}
-      {orderErrorText && (
-        <div className={classNames(styles.error, 'text', 'text_type_main-default')}>
-          При обработке заказа произошла ошибка... 😞
-          <span className={classNames('mt-2', 'text', 'text_type_main-default', 'text_color_inactive')}>{orderErrorText}</span>
-        </div>
-      )}
-      {!isLoading && !hasError() && ingredients.length > 0 && (
+      {!isLoading && ingredients.length > 0 && (
         <DndProvider backend={HTML5Backend}>
           <main className={commonStyles.content}>
             <section className={classNames(styles.contentBlock, 'mt-10')}>
